Removes statistics handler in Jetty server.
I had really just replicated what we were doing before in the old
`otj-httpserver` component, but I don't think this is necessarily
correct now.  Specifically, by overriding the handler on the server,
nothing was working.  It seems like it makes more sense to get the
handler that Spring Boot has set up, and then add our customizers to
that.  The `otj-metrics` customizer adds an `InstrumentedHandler`
in-between, so I don't think we need the `StatisticsHandler` to get nice
stats like we might have at some point in the past.