Returns Spring context-specific MBeeanServer for injection.
Jira: OTPL-1406

Instead of the static one.

We will have to be attentive to third-party libraries that try to use a
static MBeanServer.  In that circumstance, we'd just have to whip up a
connector for their MBS and the one we inject.

If they didn't support such a integration, though, then we'd have to
re-visit this approach.