Sets the main application class.
Jira: OTPL-1813

This fixes one bug and improves another circumstance.

The bug is when you use a JUnit timeout (e.g., `@Test(timeout = ...)`)
and are running a Spring application yourself (e.g.,
`OTApplication.run(...)`).  (E.g., for multi-service integration
testing.) What occurs is that the Spring main application class
automatic detection (which searches the stack trace) fails, and the main
application class is `null`.  Consequently, the application can't start
up, and you get an NPE.

The improved circumstance is also related to testing.  What occurs is
that the Spring main application class automatic detection finds that
the main class that's running is actually *not* the configuration class
that you pass in, but is instead one of the JUnit runner classes.
Consequently, the logging of how long it took for your application to
start up does *not* list your class.

This commit adds unconditional main application class setting in our
boot run wrapper.  This avoids the NPE under a timeout'ing test and the
startup time logging under test now shows the right application class.