Moves test static resources directory.
Jira: OTPL-1813

When depending on the otj-server test artifact (e.g., for the mbean
configuration), the test resources get pulled in as well.  When trying
to run another server in an integration test (e.g., discovery), the
static resources configuration can find that path instead of the one
you're actually trying to use.

Subsequent commits will update the code to test against this directory
more correctly.