Tweak more
The main problem was that we were violating original contract of otj-server. It was scanning connectors in the order of definitions in the "ot.httpserver.active-connectors" property, and if port is -1 (by default or provided) it allocated PORTx. So "boot" and "default-http" not necessarily first ones. Also "boot" and "default-http" could co-exist together. so, what I did:

1) Pre-build map of PortSelection completely scanning "ot.httpserver.active-connectors"
2) dropped PortIterator
3) portIndex is now inside PortSelector
4) Logic of PortSelector::get() slightly change
  a) First we read spring property and only if it is null or -1 we falling back (we do not want to increment counter, if value provided)
  b) For k8s, to named ports
  c) for singularity, to the ordinal ports (incrementing portIndex)
5) SpringPortSelectionPostProcessor is now simplified and providing everything to the EmbeddedJetty

As a bonus, it is printing "Port Selections" for named connectors as-well